requirements.txt
streamlit
crewai
crewai-tools
langchain
langchain-community
langchain-groq
duckduckgo-search
fastapi
uvicorn
requests
python-dotenv
pydantic